+++ integration-test/run.sh
@@ -48,12 +48,21 @@ DIST=$(cd $HERE/..; echo $PWD)
 SKELETON=$(cd $HERE/skeleton; echo $PWD)
 
 # Stackage resolver series to use
-: "${RESOLVER_SERIES:=lts-13}"
+: "${RESOLVER_SERIES:=lts-14}"
 
 # Find the latest resolver in the series to use.
 RESOLVER=$(curl -s https://www.stackage.org/download/snapshots.json | \
                jq -r '."'$RESOLVER_SERIES'"')
 echo "Using resolver: $RESOLVER"
+
+# Extra dependencies to use for the resolver
+EXTRA_DEPS_YAML=$(cd $HERE; echo $PWD)/extra-deps.$RESOLVER_SERIES
+if [ -f $EXTRA_DEPS_YAML ]
+then
+    EXTRA_DEPS=$EXTRA_DEPS_YAML
+else
+    EXTRA_DEPS=/dev/null
+fi
 
 if [ -n "$REUSE_DIR" ]
 then
@@ -84,11 +93,20 @@ fi
 cd $DIR
 
 # Copy the test files over, replacing the values
-SED="sed s!NAME!$NAME!g;s!DIST!$DIST!g;s!RESOLVER!$RESOLVER!g;s!DOCKER_DEFAULT!$DOCKER!g"
+skeleton() {
+    mkdir -p $(dirname $1)
+    sed "s!NAME!$NAME!g
+s!DIST!$DIST!g
+s!RESOLVER!$RESOLVER!g
+s!DOCKER_DEFAULT!$DOCKER!g
+/EXTRA_DEPS/{
+r$EXTRA_DEPS
+d
+}" < $SKELETON/$1 > $1
+}
 for FILE in $(find $SKELETON -type f | grep -v /\\. | sed "s!$SKELETON/!!")
 do
-    mkdir -p $(dirname $FILE)
-    $SED < $SKELETON/$FILE > $FILE
+    skeleton $FILE
 done
 
 export PATH=$(npm bin):$PATH
@@ -102,7 +120,9 @@ npm install serverless-offline
 assert_success "sls package" sls package
 
 # Test packaging without Docker
-FORCE_DOCKER=false sls package > no_docker_sls_package.txt
+# This might fail due to glibc check, ignore the failure (but still compare
+# the output)
+(FORCE_DOCKER=false sls package || true) > no_docker_sls_package.txt
 assert_success "custom variable disables Docker" \
                 grep -q "Serverless: Warning: not using Docker to build" no_docker_sls_package.txt