+++ ChangeLog.md
@@ -4,6 +4,12 @@ For the changes in v0.6.x, see this file on the corresponding branch.
 
 ## Unreleased changes
 
+* Support LTS 14.
+* Check glibc version of the resulting executable to avoid errors when running
+  (#114).
+* Use LTS 13 Docker image for building the binary to avoid depending on glibc
+  version not present on AWS.
+
 ## 0.9.1
 
 * Release a version following the proper release process.